requirements.txt
aiohttp==3.11.18
python-dotenv==1.0.1

